+++ plugins/vli/fu-vli-usbhub-rtd21xx-device.c
@@ -440,7 +440,7 @@ fu_vli_usbhub_rtd21xx_device_write_firmware(FuDevice *device,
 						    error)) {
 			g_prefix_error(error,
 				       "failed to write @0x%04x: ",
-				       fu_chunk_get_address(chk));
+				       (guint)fu_chunk_get_address(chk));
 			return FALSE;
 		}
